Rework model validation fully.
Original code was wrong, as it followed relocation,
which meant if you deployed POM that was relocating,
validation happened against effective model you were
relocating TO, and not actually deployed model.

Using latest MIMA MMR as it does exactly what we need,
and using latest MIMA as it allows to use MMR in cases
like this (no MIMA context, only in tests!).