New mojo: check-artifacts-availability (#131)
A mojo checks a list of artifacts on the given remote repository. By default it will block/wait given time (and poll), but can be used for "one pass" checks as well.

By default mojo will take a store and publisher and use those as source of artifacts to check and publisher target as remote repository, but mojo is versatile, and it can be used with properties giving comma separated list of artifacts and explicitly given remote repository as well.

Example of full manual configuration with happy outcome (all available):
https://gist.github.com/cstamas/bb54fe75e310bd4a5f884b3885c97044

Example of full manual configuration with unhappy outcome (one non existent):
https://gist.github.com/cstamas/87a2db25076cdc48762c144a8c40378a